+++ src/core/calendar/MyCalendar/MyCalendar.ts
@@ -1,0 +1,16 @@
+import { BaseCalendar } from "../BaseCalendar";
+import { ProfileStory } from "../../story/ProfileStory/ProfileStory";
+
+type Story = ProfileStory;
+
+export type MyCalendar = BaseCalendar;
+
+export const createMyCalendar = ({
+  stories = [],
+}: {
+  stories?: Story[];
+}): MyCalendar => {
+  return {
+    stories,
+  };
+};
