+++ src/redux/features/stories.ts
@@ -3,7 +3,7 @@ import { addYears, addMonths } from "date-fns";
 import {
   CommunityCollegeAfterwardsWorkingHolidayStory,
   build,
-} from "../../core/story/CommunityCollegeAfterwardsWorkingHolidayStory";
+} from "../../core/story/CommunityCollegeAfterwardsWorkingHolidayStory/CommunityCollegeAfterwardsWorkingHolidayStory";
 import { AGE_OF_START_STORY } from "../../constants/fullcalendar/options";
 import { WORKING_HOLIDAY_APPLICATION_LIMITATION_AGE } from "../../constants/visa";
 import { COMMUNITY_COLLEGE_EXAMPLE1 } from "../../constants/school";
